Changed Application File Structure.
- Install plugins named "eslint-plugin-import", "babel-plugin-module-resolver".
- Add config file named ".env" and "jsconfig.json".
- Modified config filed associated "eslint" and "babel", "webpack".
- *** Checked working well of application after change!
