+++ babel.config.js
@@ -10,4 +10,14 @@ module.exports = {
     ],
     '@babel/preset-react',
   ],
+  plugins: [
+    [
+      'module-resolver', {
+        root: ['./src'],
+        alias: {
+          '~/*': './src',
+        },
+      },
+    ],
+  ],
 };
