+++ src/__fixtures__/categories.js
