+++ src/domain/menus/component/Categories.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import styled from '@emotion/styled';
 
-import { isEmpty } from './utils';
+import { isEmpty } from 'utils/common';
 
 
 const MenuList = styled.ul`
